requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.3
decorator==5.1.1
executing==0.8.2
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
six==1.16.0
stack-data==0.1.4
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
